+++ satella/coding/concurrent/__init__.py
@@ -9,9 +9,11 @@ from .id_allocator import IDAllocator
 from .timer import Timer
 from .functions import parallel_execute, run_as_future
 from .sync_threadpool import sync_threadpool
+from .futures import Future, WrappingFuture
 
 __all__ = ['LockedDataset', 'Monitor', 'RMonitor', 'CallableGroup', 'TerminableThread',
            'MonitorDict', 'MonitorList', 'Condition', 'LockedStructure', 'AtomicNumber',
            'CallNoOftenThan', 'SingleStartThread', 'IDAllocator', 'call_in_separate_thread',
            'BogusTerminableThread', 'Timer', 'parallel_execute', 'run_as_future',
-           'sync_threadpool', 'IntervalTerminableThread']
+           'sync_threadpool', 'IntervalTerminableThread', 'Future',
+           'WrappingFuture']