+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.11.25
+
+* added `Future` and `WrappingFuture`